Renmae var.enable_health_check to var.using_load_balancer
Although var.enable_health_check is more descriptive of what setting the
variable actually does, in practice, you will want to set that only if
you are using a load balancer.

Normally, we couldn't justify such a breaking change, but we're already
(appropriately) breaking plenty in the Pull Request in which this commit
is submitted.